[REFACTOR] validate&test 도메인 별로 분리+폴더구조변경
기존의 common에 모두 있었는데
date, hour, roomkey, reseponse 로 도메인 분리함
